New system for handling collision updates
Before changing an entity's position or orientation, the function SetBookmark should be called to keep track of the entity's current state. It can be easily returned to this state by calling LoadBookmark. If the change is applied, the function UpdateCollisionChange should be called to clear the collision data at the bookmarked position and orientation, and set the new data at the new position.